Muy probablemente este roto
Todos los Ids se cambiaron a strings y aparentemente eso rompió todo el proyecto

+++ app/src/main/java/com/fsd/proyectoedson10/AddMyListActivity.kt
@@ -126,7 +126,7 @@ class AddMyListActivity : AppCompatActivity() {
                 nameList.setText(edName.text.toString())
                 val drawerLayout = AppDatabase.getDrawer()
                 drawerLayout.closeDrawers()
-                background.setBackgroundColor(db.ListDAO().selectList(db.ListDAO().selectByName(edName.text.toString()).idList).listColor.toInt())
+                background.setBackgroundColor(db.ListDAO().selectList(db.ListDAO().selectByName(edName.text.toString()).idList.toInt()).listColor.toInt())
                 true
             }
